+++ src/sparrow.h
@@ -32,7 +32,7 @@ typedef struct Object {
   char* str_val;
 
   // collections
-  int count;
+  long count;
 
   // arrays
   struct Object** array_ref;
@@ -66,8 +66,8 @@ bool is_pair(Object* cons);
 bool is_list(Object* cons);
 
 Object* make_array(size_t size);
-Object* array_at(Object* array, int index);
-void array_set(Object* array, int index, Object* value);
+Object* array_at(Object* array, size_t index);
+void array_set(Object* array, size_t index, Object* value);
 void array_print(Object* array);
 bool is_array(Object* array);
